+++ src/components/utilities/Modal.svelte
@@ -1,11 +1,18 @@
 <script lang="ts">
+    import { createEventDispatcher } from "svelte";
+
     export let title = "Edit Info"; 
     export let open = true;
+    const evt = createEventDispatcher()
+
+    const emitClickOut = () => {
+        evt('click-out')
+    }
 </script>
 
 
 
-<div role="presentation" class:open >
+<div role="presentation" class:open on:click|self={emitClickOut}>
     <div class="modal" role="dialog" aria-modal="true">
         {#if title}
             <h2>{title}</h2>
@@ -18,6 +25,7 @@
 <style lang="scss">
 div[role=presentation] {
     --_bg: var(--bg, #1b1b1b80);
+    --_z-index: var(--z-index, 5);
     display: none;
     justify-items: center;
     position: fixed;
@@ -25,10 +33,15 @@ div[role=presentation] {
     width: 100%;
     height: 100%;
     background: var(--_bg);
-    z-index: 5;
+    z-index: var(--_z-index);
+
+    h2 {
+        font-size: var(--font-size, var(--step-3));
+    }
 
     &.open {
         display: grid;
+        backdrop-filter: var(--filter, none);
     }
 }
 
@@ -36,10 +49,12 @@ div[role=presentation] {
     --_padding: 1.25rem 1.25rem 2rem 1.25rem;
     --_gap: 1.25rem;
     background: var(--clr-grey-900);
+    position: relative;
     margin-top: 5vh;
     max-height: 90vh;
     border-radius: var(--radius);
     overflow-y: auto;
+    overscroll-behavior: contain;
     // box-shadow: var(--shadow-elevation-high);
 }